# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BATIQA AI Guest Assistant API
  version: 1.0.0
  description: |
    API backend untuk asisten tamu Hotel BATIQA Pekanbaru. Menghubungkan
    tamu (via QR token) dengan pipeline AI (Gemini + fallback rule-based),
    pembuatan tiket operasional, dan dasbor staf secara real-time (SSE).

    ## Autentikasi
    - **Token tamu (QR)**: header `X-Guest-Token` atau query `?t=<token>`.
      Token dihasilkan dari `GET /api/rooms/{room}/qr`, bertanda tangan HMAC-SHA256.
    - **Staf**: header `Authorization: Bearer <token>` dari `POST /api/staff/login`.

    Kontrak lengkap dapat dilihat di UI Swagger interaktif pada `/api/docs`.
servers:
  - url: http://localhost:8080
    description: Lokal
...
paths:
...
  /api/tickets:
    post:
      tags: [Tickets]
      summary: Buat tiket permintaan layanan
      security:
        - GuestToken: []
        - StaffAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketCreateRequest'
      responses:
        '201':
          description: Tiket berhasil dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags: [Tickets]
      summary: Daftar tiket (dengan filter)
      description: |
        Staf melihat semua tiket. Tamu QR selalu dibatasi ke kamar mereka sendiri.
      security:
        - GuestToken: []
        - StaffAuth: []
      parameters:
        - name: department
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TicketStatus'
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/Priority'
        - name: room
          in: query
          schema: { type: string }
        - name: room_number
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Daftar tiket
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/tickets/{ticket_number}:
    get:
      tags: [Tickets]
      summary: Detail satu tiket
      security:
        - GuestToken: []
        - StaffAuth: []
      parameters:
        - name: ticket_number
          in: path
          required: true
          schema:
            type: string
            example: TKT-000001
      responses:
        '200':
          description: Detail tiket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tickets/{ticket_number}/status:
    patch:
      tags: [Tickets]
      summary: Perbarui status tiket (staff only)
      security:
        - StaffAuth: []
      parameters:
        - name: ticket_number
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: '#/components/schemas/TicketStatus'
      responses:
        '200':
          description: Status diperbarui
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket_number: { type: string }
                  status:
                    $ref: '#/components/schemas/TicketStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tickets/{ticket_number}/priority:
    patch:
      tags: [Tickets]
      summary: Perbarui prioritas tiket (staff only)
      security:
        - StaffAuth: []
      parameters:
        - name: ticket_number
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [priority]
              properties:
                priority:
                  $ref: '#/components/schemas/Priority'
      responses:
        '200':
          description: Prioritas diperbarui
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket_number: { type: string }
                  priority:
                    $ref: '#/components/schemas/Priority'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    GuestToken:
      type: apiKey
      in: header
      name: X-Guest-Token
      description: Token QR bertanda HMAC dari /api/rooms/{room}/qr (bisa juga via query ?t=)
    StaffAuth:
      type: http
      scheme: bearer
      bearerFormat: Token sesi staf (dari /api/staff/login)
  schemas:
...
    TicketCreateRequest:
      type: object
      required: [room_number, department, category, description, priority]
      properties:
        room_number: { type: string }
        department:
          $ref: '#/components/schemas/Department'
        category:
          type: string
          description: Intent/kategori valid (mis. TOWEL_REQUEST, AC_PROBLEM)
        description:
          type: string
          maxLength: 500
        priority:
          $ref: '#/components/schemas/Priority'
    TicketCreated:
      type: object
      properties:
        ticket_number: { type: string }
        room_number: { type: string }
        department:
          $ref: '#/components/schemas/Department'
        category: { type: string }
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/TicketStatus'
    Ticket:
      type: object
      properties:
        ticket_number: { type: string }
        room_number: { type: string }
        department:
          $ref: '#/components/schemas/Department'
        category: { type: string }
        description: { type: string }
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/TicketStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TicketDetail:
      allOf:
        - $ref: '#/components/schemas/Ticket'
        - type: object
          properties:
            resolved_at:
              type: string
              format: date-time
              nullable: true
...
    Department:
      type: string
      enum: [HOUSEKEEPING, ENGINEERING, FRONT_OFFICE]
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    TicketStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, RESOLVED, CANCELLED]
  responses:
    BadRequest:
      description: Permintaan tidak valid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Sumber daya tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }